refactor: 패키지 구조 및 이름 변경
- 미션에서 공통적으로 사용되는 네이밍(user) 준수하기 위해 member 패키지를 user로 변경하였습니다.
- 기존 패키지 구조는 계층별로 나누어져 있어서, 각 도메인과 관련된 클래스들이 흩어져 있어서
이를 한 곳에서 관리하고자 패키지 구조를 변경하였습니다.
