+++ src/main/java/com/kakao/cafe/users/domain/JDBCMemberRepository.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.domain.member;
+package com.kakao.cafe.users.domain;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
