+++ src/test/java/com/kakao/cafe/users/MemberServiceTest.java
@@ -1,7 +1,8 @@
-package com.kakao.cafe.service.member;
+package com.kakao.cafe.users;
 
-import com.kakao.cafe.domain.member.Member;
-import com.kakao.cafe.domain.member.MemberRepository;
+import com.kakao.cafe.users.domain.Member;
+import com.kakao.cafe.users.domain.MemberRepository;
+import com.kakao.cafe.users.MemberService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
